Update README.md
Changed PREVIOUS MEETINGS to RECENT MEETINGS, and reduced table to only reflect the 5 most recent meetings.  This improves readability of the README text, and assists with wayfinding to the most relevant materials.  Included links to the meeting notes file, as well as additional  link to the Filecoin YouTube channel. 

Updated the WHO CAN ATTEND section to include information about adding core devs, adding observers, and removing a core dev.  This does not reflect any real process change, but rather makes explicit a previously implicit workflow.  Also included a reference to Coraline Ada Ehmke's Post-Meritocracy manifesto as a definition of the merit principles upon which a requested Core Dev ought to be considered. 

Updates MEETING SCHEDULING to reflect the current schedule. 

Updates AGENDA ITEMS to be more current with what is accepted as a meeting item. 

Updates WHO MANAGES THE MEETING to be current. 

Removed configuration instructions for livestreaming, since this is a closed meeting.